#382: Refactored the Vite plugin templates to use the domain folder instead of shared
