+++ packages/create-jitar/templates/lit/segments/default.segment.json
@@ -1,3 +1,3 @@
 {
-    "./shared/sayHello": { "sayHello": { "access": "public" } }
+    "./domain/sayHello": { "sayHello": { "access": "public" } }
 }
