+++ packages/create-jitar/templates/lit/src/domain/sayHello.ts
