+++ packages/create-jitar/templates/react/src/App.tsx
@@ -4,7 +4,7 @@ import './App.css'
 import reactLogo from './assets/react.svg'
 import jitarLogo from './assets/jitar.svg'
 
-import { sayHello } from './shared/sayHello'
+import { sayHello } from './domain/sayHello'
 
 function App()
 {
