+++ packages/create-jitar/templates/react/src/domain/sayHello.ts
