+++ packages/create-jitar/templates/react/vite.config.ts
@@ -5,6 +5,6 @@ import jitar from '@jitar/plugin-vite'
 export default defineConfig({
   plugins: [
     react(),
-    jitar('src', 'shared', 'http://localhost:3000')
+    jitar('src', 'domain', 'http://localhost:3000')
   ]
 })
