+++ packages/create-jitar/templates/solid/src/domain/sayHello.ts
