+++ packages/create-jitar/templates/solid/vite.config.ts
@@ -5,7 +5,7 @@ import jitar from '@jitar/plugin-vite'
 export default defineConfig({
   plugins: [
     solidPlugin(),
-    jitar('src', 'shared', 'http://localhost:3000')
+    jitar('src', 'domain', 'http://localhost:3000')
   ],
   build: {
     target: 'esnext',
