+++ packages/create-jitar/templates/svelte/src/domain/sayHello.ts
