+++ packages/create-jitar/templates/svelte/vite.config.ts
@@ -5,6 +5,6 @@ import jitar from '@jitar/plugin-vite'
 export default defineConfig({
   plugins: [
     svelte(),
-    jitar('src', 'shared', 'http://localhost:3000')
+    jitar('src', 'domain', 'http://localhost:3000')
   ],
 });
