+++ packages/create-jitar/templates/vue/src/domain/sayHello.ts
