+++ packages/create-jitar/templates/vue/vite.config.ts
@@ -5,5 +5,5 @@ import jitar from '@jitar/plugin-vite'
 export default defineConfig({
   plugins: [
     vue(),
-    jitar('src', 'shared', 'http://localhost:3000')],
+    jitar('src', 'domain', 'http://localhost:3000')],
 })
